Split server and app into two files
